refactor(seances): SeanceManagement sous 300 lignes — éclatement complet (H6)
Éclatement du god-component (coordinateur) en composable + 5 sous-composants,
CSS VERBATIM distribué (chaque @media replacé avec sa cible) :
- SeanceManagement.vue 1051→205 (orchestrateur)
- useSeanceManagement.js (213) UI state + actions visio + routage calendrier
  (compose le useCoordinatorSeances existant, consommé sans modification)
- CoordinatorSeanceHeader.vue (126) en-tête + bascule liste/calendrier (v-model)
- CoordinatorSeanceFilters.vue (130) filtres période/enseignant/classe
- CoordinatorSeanceCard.vue (225) carte infos + toggle visio (+ panneau)
- CoordinatorVisioPanel.vue (229) panneau visio (participants/rejoindre/attente)
- CoordinatorSeanceStats.vue (81) total/visios/taux
Tests: montage + composable + sous-composants = 13 verts. vite build vert.
.btn-icon dupliqué verbatim (carte + panneau) car partagé → scoping per-composant.
